+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 <!-- Add new, unreleased items here. -->
+- Use Github API Pagination to traverse pages
 
 ## v2.1.0 [01-10-2018]
 - Made `Workspace.determineGitHubRepos()` method public.